archive/MOSB_Logistics_Dashboard_Phase3_1_Pack/mosb_logistics_dashboard_phase3_1_pack/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
duckdb>=0.9.0
cachetools>=5.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
